Revert "move OrderByDefinition to out from FilteredRecordRetrievalOptions so it's not a nested type anymore (requested during API review)"
This reverts commit 10c224df2ad1c8665ffdc41cdb286e0b1c09de5b.
